+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## Unreleased - ???
 - Update meson build script to fix bug on Debian's version of meson
 - Add `xprint`, `xprin`, `xprintf`, and `xprinf`.
+- `net/write` now returns an error message if write fails.
+- Fix issue with SIGPIPE on macOS and BSDs.
 
 ## 1.11.3 - 2020-08-03
 - Add `JANET_HASHSEED` environment variable when `JANET_PRF` is enabled.